requirements.txt
allure-pytest==2.9.45
selenium==4.4.3
pytest==7.1.3
